Remove edited_by_human?
This method is not currently used. I haven't been able to determine how
long ago it became unused.

The intention was to allow different treatment of mappings edited by a
human versus mappings imported and not edited by a human [1], but I can
not find evidence of this different treatment in the current code.

[1] https://github.com/alphagov/transition/commit/16a8d217f8853db5dc40433336c5d7132de82534