Bring over the scripts from transition-config
Various scripts to analyse queries, clean up mappings, get information
about a site.

Some of them are referenced from the documentation on adding a new site
to transition, so the easiest course of action for now is to bring them
into this repo and update the documentation.

Some of the scripts were specific for working with the YAML files in
transition-config, so we haven't brought them over. These are:
- `known_domains.sh`
- `update_webarchive.sh`

Original location: https://github.com/alphagov/transition-config/tree/main/tools