Remove console message with out-of-date advice
We've recently retired the transition-config project and no longer
import anything from it.

I'm not all that familiar with hits. I know that they were formerly
imported from transition-config and that later on that was replaced by
the use of S3. So I'm assuming that the reference to a directory relates
to actual copies of transition-config's flat files. Which is obviously
no longer relevant, either.