Stub responses from Organisations API in specs
This removes the need to import organisations directly from a YAML file,
so the code that supported the functionality is being removed here too.

(But since it still uses the YAML fixtures as the source for the stub,
the fixture files themselves are staying.)

The result is less code paths and less special cases, plus more of
Transition's production code being exercised in its test suite.