Remove the revert Rake tasks
* Delete `import:revert:sites` and `import:revert:hosts` Rake tasks
* Delete supporting code for `import:revert:sites` and `import:revert:hosts`
* Delete `import:revert_entirely_unsafe` Rake task
* But leave the supporting code for `import:revert_entirely_unsafe` in place

`import:revert:sites` and `import:revert:hosts` were used to delete
sites that didn't have any mappings or hits. We no longer use these.

We've added site deletion to the Transition app's web interface (with
plenty of warnings and papertrail audit trails) and that uses "revert
entirely" under the hood. But we don't want it to be used from a rake
task anymore!

For anybody wondering about the ominously named "revert entirely
unsafe": the "entire" qualifier applies to the revert, not to the lack
of safety.