Drop spec's dependency on Sites YAML fixtures
This spec loaded 8 fixtures from `spec/fixtures/sites/someyaml/` but
only directly referred to 2 (`ago.yml` and `bis.yml`).

I've used factories to reproduce (what I believe to be) the relevant
details of those 2 fixtures, and updated the global database count
assertions to reflect the loss of the other 6 Sites (and their 14
Hosts).

Importing `Site`s from YAML files is a hangover from Transition's use of
the (recently retired) transition-config project. It isn't unreasonable
to use YAML fixtures for specs (!) but this specific spec is one of the
last remaining users of this Site-importing subsystem and dropping its
dependency on it might just allow us to remove the lot.