Remove scenario that emulates transition-config
The transition-config project was recently retired and the Transition
app no longer imports data from it.

This scenario seems to be intended to emulate the no-longer-existent
interaction between Transition and transition-config. (It's barely
changed since [its introduction 10 years
ago](5fdb7207ce8cb2086c70b225d48d85cb3cf74fa8) and has always created
the Mapping using a factory.)

I'm not familiar enough with Transition yet to say for sure that there's
no new process analogous to the old import from transition-config that
could create a Mapping with no history and that would be worth testing,
but I'm betting on there not being one and so I'm just deleting this.