Move class method into CanonicalFilter class
Currently, the `canonical_filter` class method is located in the
`View::Mappings` module. This is available in the `Filter` class (which is also
located in `View::Mappings`) because the entire
`view/mappings/canonical_filter` file is loaded manually with `require`.

However, in order to use the Rails 7.1defaults, this require must be removed as
we can't manually require files that are managed by the autoloader. This means
that `canonical_filter` will no longer be available in the `Filter` class.

This moves the functionality of `canonical_filter` into a factory method on the
`CanonicalFilter` class.