Update class names with abbreviations
Using the Rails 7.1 defaults, we see an errors where file names are not what is
expected by the autoloader.

All of the examples here are due to the capitalisation of abbreviations.

This aims for some consistency (all acronyms are capitalised) by either
updating the Zetwerk config or by updating the class name, depending on whether
Zeitwork is natively aware of the abbreviation.