Use site ID in URLs instead of abbreviation
This will allow us to deprecate the need for an abbrevation for sites,
by using the site's ID in the URL instead of a user input abbreviation.

It retains backward compatibility, so abbreviation can still be used as
a slug (e.g. if a user has bookmarked an existing page).