Sanitize and cleanup invalid UTF-8 characters
It is possible for users to incldue invalid or null byte characters in
their requests.

Therefore adding middleware to sanitise request URIs, headers and
cookies before the request is processed by Rails.

Also adds a test to reproduce `ArgumentError` exception which as a
result of an automated scan being run by Cyber Security which includes a
null byte in the request (we believe it is in the session cookie).