Increase dependabot PR limit for Ruby gems
Dependabot has a default limit of 5 for the number of pull requests it
will open on a repo [1]. We have missed some updates this week, since
Dependabot has been limited to the number of PRs it can open.

Therefore increasing the limit to 20, as we are unlikely to exceed this.

1: https://docs.github.com/en/code-security/dependabot/working-with-dependabot/troubleshooting-dependabot-errors#dependabot-cannot-open-any-more-pull-requests