Use Capybara RSpec matcher instead of expect/eql
We're seeing this expectation fail fairly consistently at the minute. In those
failures the actual URI it's got has a `return_path` query param, so I'm
thinking that the expectation is just running faster than the redirect.

As far as I know, the Capybara version should wait as it would for any
other page content-related expectation.