Wait for page to load after jump to site/mapping
I don't really understand the cause of it, but some scenarios in
`site.feature` have been raising exceptions in CI from the
`Then I should see the header "..."` step when it immediately follows
the `And I jump to the site or mapping "..."` step.

The exceptions are being reported like:
```
unknown error: unhandled inspector error: {"code":-32000,"message":"Node with given id does not belong to the document"}
  (Session info: chrome=134.0.6998.35) (Selenium::WebDriver::Error::UnknownError)
```

It doesn't look like a clear-cut case of not waiting for the page to
load. If I've read the error message correctly, it sounds like our
matcher successfully matches a node on the previous page (or at least
not the page we're heading towards??) and then the target page finishes
loading before the expectation is executed against that matched node.
So, this is an attempt to wait for the page to load before even moving
onto that next scenario step and its matcher/expectation. It sounds like
a long shot, but I'm running out of ideas.