Wait for page to load after filtering mappings
This change is based on the same assessment as the previous commit's.

In the case of filtering mappings, the previous current_path and the new
current_path are pretty much identical and therefore tricky to match
against. The difference is a query param, but it seems like it can have
any one of a number of names based on the filter type being used.